requirements.txt
python-dateutil~=2.8
tweepy~=3.9

# python-dotenv~=0.14